baby-care_google_project/babycare_agent_app/requirements.txt
google_adk==1.19.0
dotenv==0.9.9
